backend/requirements.txt
# Web Framework
fastapi==0.134.0
uvicorn==0.41.0
python-multipart==0.0.22
slowapi==0.1.9

# Config
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.1

# LLM (Bedrock primary, Groq fallback — both via openai client)
openai==2.24.0
instructor==1.14.5
google-genai==0.4.0  # Legacy Gemini support

# Templates
jinja2==3.1.6

# Memory (Redis)
redis==7.2.1

# AWS (Bedrock LLM, Textract OCR, S3 Storage)
boto3==1.42.59

# Async HTTP & File I/O
aiohttp==3.13.3
aiofiles==25.1.0

# PDF Generation
weasyprint>=62.0
pypdf>=5.0

# Supabase (Shelter DB)
supabase>=2.0
thefuzz>=0.22

# RAG Pipeline
cohere>=5.0
chromadb>=1.0
beautifulsoup4>=4.0
